Add redirect for our golang packages
Now that we're moving everything in preparation for open-sourcing it,
we want to ensure that we don't have to do it again. The solution is
golang "vanity packages" where the go tool can follow a <meta> tag
from a vanity name (e.g., epic-gateway.org) to its repo (e.g.,
github.com/epic-gateway).

To see the effect of this commit, go to /marin3r and that page will
have two meta tags: go-import and go-source.

This was a helpful resource:
https://blog.jbowen.dev/2020/07/using-go-vanity-urls-with-hugo/

Signed-off-by: Toby Cabot <toby@acnodal.io>